✨feat(rbac): 新增角色和菜单关联（权限）接口

+++ cool-module-system/cool-module-system-infrastructure/src/main/java/com/xgblack/cool/module/system/gateway/PermissionGatewayImpl.java
@@ -3,8 +3,11 @@ package com.xgblack.cool.module.system.gateway;
 import com.mybatisflex.core.query.QueryChain;
 import com.mybatisflex.core.query.QueryWrapper;
 import com.xgblack.cool.module.system.domain.gateway.PermissionGateway;
+import com.xgblack.cool.module.system.gateway.database.dataobject.RoleMenuDO;
 import com.xgblack.cool.module.system.gateway.database.dataobject.UserRoleDO;
+import com.xgblack.cool.module.system.gateway.database.dataobject.table.RoleMenuTableDef;
 import com.xgblack.cool.module.system.gateway.database.dataobject.table.UserRoleTableDef;
+import com.xgblack.cool.module.system.gateway.database.mapper.RoleMenuMapper;
 import com.xgblack.cool.module.system.gateway.database.mapper.UserRoleMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -26,6 +29,8 @@ import java.util.Set;
 public class PermissionGatewayImpl implements PermissionGateway {
 
     private final UserRoleMapper userRoleMapper;
+
+    private final RoleMenuMapper roleMenuMapper;
 
 
     @Override
@@ -60,7 +65,45 @@ public class PermissionGatewayImpl implements PermissionGateway {
         }
         QueryWrapper query = QueryWrapper.create()
                 .from(UserRoleTableDef.USER_ROLE)
+                .and(UserRoleTableDef.USER_ROLE.USER_ID.eq(userId))
                 .and(UserRoleTableDef.USER_ROLE.ROLE_ID.in(roleIds));
+        userRoleMapper.deleteByQuery(query);
+    }
+
+    @Override
+    public Set<Long> getMenuIdsByRoleId(Long roleId) {
+        List<Long> menuIds = QueryChain.of(roleMenuMapper)
+                .select(RoleMenuTableDef.ROLE_MENU.MENU_ID)
+                .from(RoleMenuTableDef.ROLE_MENU)
+                .and(RoleMenuTableDef.ROLE_MENU.ROLE_ID.eq(roleId))
+                .objListAs(Long.class);
+        if (CollUtil.isEmpty(menuIds)) {
+            return SetUtil.empty();
+        }
+        return SetUtil.of(menuIds);
+    }
+
+    @Override
+    public void insertRoleMenu(Long roleId, Set<Long> menuIds) {
+        if (CollUtil.isEmpty(menuIds)) {
+            return;
+        }
+        List<RoleMenuDO> list = ListUtil.empty();
+        for (Long menuId : menuIds) {
+            list.add(new RoleMenuDO().setRoleId(roleId).setMenuId(menuId));
+        }
+        roleMenuMapper.insertBatch(list);
+    }
+
+    @Override
+    public void deleteRoleMenu(Long roleId, Set<Long> menuIds) {
+        if (CollUtil.isEmpty(menuIds)) {
+            return;
+        }
+        QueryWrapper query = QueryWrapper.create()
+                .from(RoleMenuTableDef.ROLE_MENU)
+                .and(RoleMenuTableDef.ROLE_MENU.ROLE_ID.eq(roleId))
+                .and(RoleMenuTableDef.ROLE_MENU.MENU_ID.in(menuIds));
         userRoleMapper.deleteByQuery(query);
     }
